# Optimized narrow mmt4d tile functions (for small M0)