Data-tiling properly selects narrow power-of-two tile dimensions when the default tile size would be greater than the generic tile size, resulting in relatively large padding. For example, then the input matrix is a vector (M0 == 1), it would be unfortunate to pad that to 8 or 16.

When using microkernels, this however resulted in falling off of the fast path, so far only covering the generic tile size (such as 8x8 or 16x16) and not the resulting narrow tile sizes.

This PR adds those missing ukernel fast code paths for small M0.  Nothing is done about small N0 cases; those should be handled in a future PR by transposing the problem to reduce it to a small-M0 case. There is a basic dissymetry between M0 and N0 here, which is that the tiles are row-major not colum-major, so a small M0 case simply means a truncated tile, while a smaller N0 case would require a completely different tile layout. There is no need to embark on that if we can instead reduce generality by transposition.

# Benchmarks on x86-64, AMD Zen4 (Ryzen 7940 HS):

| Benchmark                                   | Gop/s before | Gop/s after | Speedup x |
| ------------------------------------------- | ------------ | ----------- | --------- |
| BM_mmt4d_f32f32f32_tile_1x8x1_avx2_fma      | 0.9          | 24.3        | 27.8      |
| BM_mmt4d_f32f32f32_tile_2x8x1_avx2_fma      | 0.9          | 42.7        | 49.2      |
| BM_mmt4d_f32f32f32_tile_4x8x1_avx2_fma      | 1.4          | 80.8        | 57.4      |
| BM_mmt4d_f32f32f32_tile_8x8x1_avx2_fma      | 39.3         | 130.3       | 3.3       |
| BM_mmt4d_f32f32f32_tile_1x16x1_avx512_base  | 1.0          | 42.4        | 43.1      |
| BM_mmt4d_f32f32f32_tile_2x16x1_avx512_base  | 1.0          | 84.9        | 82.0      |
| BM_mmt4d_f32f32f32_tile_4x16x1_avx512_base  | 1.0          | 138.7       | 142.8     |
| BM_mmt4d_f32f32f32_tile_8x16x1_avx512_base  | 1.1          | 163.5       | 150.7     |
| BM_mmt4d_f32f32f32_tile_16x16x1_avx512_base | 89.4         | 163.7       | 1.8       |
| BM_mmt4d_f16f16f32_tile_1x8x1_avx2_fma      | 0.2          | 21.3        | 110.3     |
| BM_mmt4d_f16f16f32_tile_2x8x1_avx2_fma      | 0.2          | 40.1        | 204.5     |
| BM_mmt4d_f16f16f32_tile_4x8x1_avx2_fma      | 0.2          | 72.6        | 367.8     |
| BM_mmt4d_f16f16f32_tile_8x8x1_avx2_fma      | 23.2         | 79.2        | 3.4       |
| BM_mmt4d_f16f16f32_tile_1x16x1_avx512_base  | 0.2          | 37.5        | 171.0     |
| BM_mmt4d_f16f16f32_tile_2x16x1_avx512_base  | 0.2          | 63.2        | 256.5     |
| BM_mmt4d_f16f16f32_tile_4x16x1_avx512_base  | 0.2          | 72.5        | 293.6     |
| BM_mmt4d_f16f16f32_tile_8x16x1_avx512_base  | 0.2          | 78.1        | 315.9     |
| BM_mmt4d_f16f16f32_tile_16x16x1_avx512_base | 80.6         | 81.3        | 1.0       |
| BM_mmt4d_f16f16f16_tile_1x8x1_avx2_fma      | 0.1          | 21.2        | 167.6     |
| BM_mmt4d_f16f16f16_tile_2x8x1_avx2_fma      | 0.1          | 40.1        | 315.9     |
| BM_mmt4d_f16f16f16_tile_4x8x1_avx2_fma      | 0.1          | 72.5        | 571.3     |
| BM_mmt4d_f16f16f16_tile_8x8x1_avx2_fma      | 78.9         | 79.4        | 1.0       |
| BM_mmt4d_f16f16f16_tile_1x16x1_avx512_base  | 0.1          | 37.7        | 296.5     |
| BM_mmt4d_f16f16f16_tile_2x16x1_avx512_base  | 0.1          | 63.2        | 495.8     |
| BM_mmt4d_f16f16f16_tile_4x16x1_avx512_base  | 0.1          | 72.6        | 569.9     |
| BM_mmt4d_f16f16f16_tile_8x16x1_avx512_base  | 0.1          | 78.1        | 612.6     |
| BM_mmt4d_f16f16f16_tile_16x16x1_avx512_base | 80.9         | 81.0        | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x8x2_avx2_fma        | 2.8          | 56.5        | 20.1      |
| BM_mmt4d_i8i8i32_tile_2x8x2_avx2_fma        | 3.1          | 76.7        | 24.6      |
| BM_mmt4d_i8i8i32_tile_4x8x2_avx2_fma        | 3.2          | 91.6        | 28.5      |
| BM_mmt4d_i8i8i32_tile_8x8x2_avx2_fma        | 181.1        | 183.0       | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x16x2_avx512_base    | 3.2          | 37.3        | 11.5      |
| BM_mmt4d_i8i8i32_tile_2x16x2_avx512_base    | 3.4          | 48.8        | 14.5      |
| BM_mmt4d_i8i8i32_tile_4x16x2_avx512_base    | 3.5          | 58.7        | 16.9      |
| BM_mmt4d_i8i8i32_tile_8x16x2_avx512_base    | 3.5          | 117.4       | 33.3      |
| BM_mmt4d_i8i8i32_tile_16x16x2_avx512_base   | 230.1        | 228.8       | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x16x2_avx512_vnni    | 3.2          | 50.2        | 15.5      |
| BM_mmt4d_i8i8i32_tile_2x16x2_avx512_vnni    | 3.4          | 68.6        | 20.4      |
| BM_mmt4d_i8i8i32_tile_4x16x2_avx512_vnni    | 3.5          | 82.2        | 23.7      |
| BM_mmt4d_i8i8i32_tile_8x16x2_avx512_vnni    | 3.5          | 163.3       | 46.4      |
| BM_mmt4d_i8i8i32_tile_16x16x2_avx512_vnni   | 323.0        | 323.9       | 1.0       |

# Benchmarks on arm64, Cortex-X2:

| Benchmark                             | Gop/s before | Gop/s after | Speedup x |
| ------------------------------------- | ------------ | ----------- | --------- |
| BM_mmt4d_f32f32f32_tile_1x8x1         | 1.4          | 23.2        | 17.0      |
| BM_mmt4d_f32f32f32_tile_2x8x1         | 1.4          | 45.5        | 33.0      |
| BM_mmt4d_f32f32f32_tile_4x8x1         | 1.4          | 71.4        | 51.4      |
| BM_mmt4d_f32f32f32_tile_8x8x1         | 92.0         | 91.9        | 1.0       |
| BM_mmt4d_f16f16f32_tile_1x8x1         | 0.6          | 18.7        | 30.8      |
| BM_mmt4d_f16f16f32_tile_2x8x1         | 0.6          | 24.2        | 39.6      |
| BM_mmt4d_f16f16f32_tile_4x8x1         | 0.6          | 36.6        | 62.4      |
| BM_mmt4d_f16f16f32_tile_8x8x1         | 47.1         | 47.2        | 1.0       |
| BM_mmt4d_f16f16f32_tile_1x8x1_fp16fml | 0.6          | 23.6        | 38.9      |
| BM_mmt4d_f16f16f32_tile_2x8x1_fp16fml | 0.6          | 40.1        | 65.6      |
| BM_mmt4d_f16f16f32_tile_4x8x1_fp16fml | 0.6          | 59.2        | 101.0     |
| BM_mmt4d_f16f16f32_tile_8x8x1_fp16fml | 92.2         | 71.8        | 0.8       |
| BM_mmt4d_f16f16f16_tile_1x8x1         | 0.3          | 18.6        | 66.8      |
| BM_mmt4d_f16f16f16_tile_2x8x1         | 0.3          | 24.2        | 85.0      |
| BM_mmt4d_f16f16f16_tile_4x8x1         | 0.3          | 36.2        | 129.0     |
| BM_mmt4d_f16f16f16_tile_8x8x1         | 47.0         | 47.0        | 1.0       |
| BM_mmt4d_f16f16f16_tile_1x8x1_fp16    | 0.3          | 24.1        | 86.3      |
| BM_mmt4d_f16f16f16_tile_2x8x1_fp16    | 0.3          | 46.4        | 162.8     |
| BM_mmt4d_f16f16f16_tile_4x8x1_fp16    | 0.3          | 86.0        | 306.1     |
| BM_mmt4d_f16f16f16_tile_8x8x1_fp16    | 135.8        | 139.6       | 1.0       |
| BM_mmt4d_bf16bf16f32_tile_1x8x4_bf16  | 0.7          | 62.1        | 88.8      |
| BM_mmt4d_bf16bf16f32_tile_2x8x4_bf16  | 0.7          | 123.5       | 176.5     |
| BM_mmt4d_bf16bf16f32_tile_4x8x4_bf16  | 0.7          | 224.9       | 325.1     |
| BM_mmt4d_bf16bf16f32_tile_8x8x4_bf16  | 294.8        | 295.2       | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x8x1           | 1.5          | 20.1        | 13.6      |
| BM_mmt4d_i8i8i32_tile_2x8x1           | 1.5          | 20.4        | 13.3      |
| BM_mmt4d_i8i8i32_tile_4x8x1           | 1.6          | 21.0        | 13.5      |
| BM_mmt4d_i8i8i32_tile_8x8x1           | 47.0         | 47.0        | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x8x4_dotprod   | 3.1          | 129.0       | 41.4      |
| BM_mmt4d_i8i8i32_tile_2x8x4_dotprod   | 3.1          | 174.8       | 56.0      |
| BM_mmt4d_i8i8i32_tile_4x8x4_dotprod   | 3.0          | 357.1       | 117.9     |
| BM_mmt4d_i8i8i32_tile_8x8x4_dotprod   | 367.2        | 367.5       | 1.0       |
| BM_mmt4d_i8i8i32_tile_1x8x8_i8mm      | 8.6          | 174.1       | 20.1      |
| BM_mmt4d_i8i8i32_tile_2x8x8_i8mm      | 8.8          | 359.9       | 41.0      |
| BM_mmt4d_i8i8i32_tile_4x8x8_i8mm      | 8.7          | 488.9       | 56.0      |
| BM_mmt4d_i8i8i32_tile_8x8x8_i8mm      | 594.1        | 593.4       | 1.0       |