@bjacob 2023-09-19
Before I'll mark this ready for review, I'll solve that clang-format issue... I inlined some code, resulting in nested `#if` within `if()`, exposing differences in behavior between my clang-format and the CI's clang-format. I'll undo that inlining, that's best for readability anyway.

Oh, looks like I need to look into a GCC compile error too.