@bjacob 2023-09-19
> only a 600x improvement? :shakes head: ;P

As always with 600x improvements, it's all about bad baselines :-)