@bjacob 2023-09-20
ok @benvanik this is ready for review. summary:

* Something about the `#if`'s here is formatted differently across different versions of `clang-format` so I had to just copy and paste the diff generated by the CI.
* Leaving out old GCC (GCC < 12) from fast x86 code paths because they have different x86 intrinsics.