@stellaraccident 2023-09-20
Nice! This brings some nostalgia: I think working with you to implement these fast paths in tfmini was the first time we started working together. Full circle!